Do not generate @NotNull for notInPayload properties
Properties which are defined as not being part of the payload are no
longer annotated with @NotNull in the generated POJOs. This was
unnecessary and inconsistent.
The generated class for an Entity with a Property which is not part of
the payload contains a constructor which does not include this Property,
used for deserialization. Hence the generated class contained a member
variable annotated with @NotNull which is not set in the constructor.

fixes #42