Fix code generation for Entity-Enum with langString
When an Entity with a language string property is used as the data type
of an Enum, the Enum constants were not generated correctly. In the
generation of the Enum constant a language string was treated in the
same manner as a string, causing an exception during the code
generation.

fixes #24