Optional header for generated POJOs
The default header containing an incorrect copyright notice is removed
from the POJO generation. Instead it is now possible to provide a custom
header for the generated classes. Providing the header is optional.

The header may be provided via a Velocity template macro defined in a
Velocity template file. The path to the file as well as its name, may be
passed as arguments to the code generation. The name of the macro must
be 'copyrightHeader'.

fixes #59