Prepare Meta Model Resource Resolver for next BAMM
The next BAMM version contains a bug fix for the Locale Constraint
shape. The fix includes a generated javascript file which contains the
BCP 47 language tags. In order to load this script file, the
SdsAspectMetaModelResourceResolver is updated to handle the new
script subfolder in the resources of the BAMM project.

fixes #84