Add new test suite structure
The core idea is that the same set of tests (in SharedTests.java) are first
regularly executed as unit tests by running the CLI main function using Maven
Surefire and creating a code coverage report. Then the executable jar is built
and the tests run as integration tests using Maven Failsafe, testing the jar.
Then the native binary is built from the executable jar and the tests are run a
third time, this time testing the native binary.