Handle models with any number of Aspects in artifact generators
This adjusts the constructors of the document and code generators so that the
Aspect they intend to work on must be passed as an explicit argument. For that
the `AspectContext` is introduced which wraps a `VersionedModel` and a single
`Aspect` that is contained in the model. The task to determine the Aspect from
the loaded model is delegated to the caller (here bamm-cli and Maven plugin).