Extract handling for properties to allow more control
Property handling is now done through an indirection which
allows to decide per property or property type how it should
be handled.
For the first this is now used when writing submodels with
data, where serialized LangStrings are given, which require
a special handling, as the payload structure is different from
the metamodel structure.